requirements-cloud.txt
# Cloud orchestrator only. Local browser, audio, and cursor dependencies stay on the client machine.
google-adk==1.25.1
google-genai>=1.29.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.1
websockets==15.0.1
